Reorganized some modules to bring app in line with others.
